Changed how subpixel rendering works.

+++ river/scene/fonts/glyph-object.hpp
@@ -15,14 +15,14 @@ namespace River
 	private:
 		Glyph *glyph;
 		color_t color;
-		GLfloat offset;
+		uint8_t offset;
 	public:
 		GlyphObject();
 		~GlyphObject();
 
 		SimpleEntry<GlyphObject> text_entry;
 
-		void set_glyph(Glyph *glyph, int subpixel_offset, color_t color);
+		void set_glyph(Glyph *glyph, uint8_t subpixel_offset, color_t color);
 		
 		void render();
 		static void render_key(GlyphCache *cache);
